+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/util/ConfigurationMetadataDescriptor.java
@@ -5,7 +5,6 @@ import java.util.Properties;
 import org.hibernate.boot.Metadata;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.tool.api.metadata.MetadataDescriptor;
-import org.hibernate.tool.orm.jbt.util.MetadataHelper;
 
 public class ConfigurationMetadataDescriptor implements MetadataDescriptor {
