+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/util/MetadataHelper.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.orm.jbt.util;
+package org.hibernate.tool.orm.jbt.internal.util;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
