+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/util/NativeConfiguration.java
@@ -20,6 +20,7 @@ import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Table;
 import org.hibernate.tool.api.reveng.RevengStrategy;
 import org.hibernate.tool.orm.jbt.internal.util.ExtendedConfiguration;
+import org.hibernate.tool.orm.jbt.internal.util.MetadataHelper;
 import org.w3c.dom.Document;
 import org.xml.sax.EntityResolver;
